CertificateGen/cert/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.14
numpy==1.24.4
opencv-python==4.10.0.84
pandas==2.0.3
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1

